refactor to extract stdlib portion from deimos_shared
