+++ software/deimos/src/calcs/constant.rs
