+++ software/deimos/src/calcs/inverse_affine.rs
