+++ software/deimos/src/calcs/rtd_pt100.rs
