+++ software/deimos/src/calcs/sin.rs
