+++ software/deimos/src/calcs/tc_ktype.rs
